requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
attrs==21.2.0
    # via
    #   jsonschema
    #   pytest
awscli==1.20.19
    # via -r requirements.in
azure-cognitiveservices-vision-computervision==0.5.0
    # via -r requirements.in
azure-common==1.1.27
    # via
    #   -r requirements.in
    #   azure-cognitiveservices-vision-computervision
    #   azure-storage-blob
    #   azure-storage-common
azure-storage-blob==2.1.0
    # via -r requirements.in
azure-storage-common==2.1.0
    # via
    #   -r requirements.in
    #   azure-storage-blob
bcrypt==3.2.0
    # via paramiko
beautifulsoup4==4.9.3
    # via
    #   -r requirements.in
    #   webtest
blinker==1.4
    # via elastic-apm
boto3==1.18.19
    # via
    #   -r requirements.in
    #   credstash
    #   moto
botocore==1.21.19
    # via
    #   awscli
    #   boto3
    #   moto
    #   s3transfer
cachetools==4.2.2
    # via
    #   -r requirements.in
    #   google-auth
certifi==2021.5.30
    # via
    #   elastic-apm
    #   msrest
    #   requests
cffi==1.14.6
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
chardet==4.0.0
    # via pdfminer.six
charset-normalizer==2.0.4
    # via requests
click==7.1.2
    # via
    #   clickclick
    #   flask
clickclick==20.10.2
    # via connexion
colorama==0.4.3
    # via awscli
coloredlogs==15.0.1
    # via -r requirements.in
connexion[swagger-ui]==2.9.0
    # via -r requirements.in
credstash==1.17.1
    # via -r requirements.in
cryptography==3.4.7
    # via
    #   azure-storage-common
    #   credstash
    #   moto
    #   paramiko
    #   pdfminer.six
    #   python-jose
django==2.2.24
    # via -r requirements.in
djangorestframework==3.9.4
    # via -r requirements.in
docutils==0.15.2
    # via awscli
ecdsa==0.17.0
    # via python-jose
elastic-apm[flask]==6.3.3
    # via -r requirements.in
fakeredis[lua]==1.5.2
    # via -r requirements.in
fhirclient==1.0.6
    # via -r requirements.in
flask==1.1.4
    # via connexion
fuzzywuzzy==0.18.0
    # via -r requirements.in
gitdb==4.0.7
    # via gitpython
gitpython==3.1.18
    # via -r requirements.in
google-api-core[grpc]==1.31.1
    # via google-cloud-vision
google-auth==1.34.0
    # via google-api-core
google-cloud-vision==1.0.0
    # via -r requirements.in
googleapis-common-protos==1.53.0
    # via
    #   -r requirements.in
    #   google-api-core
grpcio==1.39.0
    # via google-api-core
humanfriendly==9.2
    # via coloredlogs
idna==3.2
    # via requests
ijson==3.1.4
    # via -r requirements.in
importlib-metadata==4.6.3
    # via moto
inflection==0.5.1
    # via connexion
iniconfig==1.1.1
    # via pytest
isodate==0.6.0
    # via
    #   fhirclient
    #   msrest
    #   openapi-schema-validator
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via
    #   flask
    #   moto
    #   swagger-ui-bundle
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
json2html==1.3.0
    # via -r requirements.in
jsonschema==3.2.0
    # via
    #   connexion
    #   openapi-schema-validator
    #   openapi-spec-validator
lupa==1.9
    # via fakeredis
lxml==4.6.3
    # via -r requirements.in
markupsafe==2.0.1
    # via
    #   jinja2
    #   moto
more-itertools==8.8.0
    # via moto
moto==2.2.2
    # via -r requirements.in
msrest==0.6.21
    # via azure-cognitiveservices-vision-computervision
nltk==3.4.5
    # via -r requirements.in
numpy==1.18.5
    # via
    #   -r requirements.in
    #   pandas
oauthlib==3.1.1
    # via requests-oauthlib
openapi-schema-validator==0.1.5
    # via openapi-spec-validator
openapi-spec-validator==0.3.1
    # via connexion
packaging==21.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   pytest
pandas==1.1.5
    # via -r requirements.in
paramiko==2.7.2
    # via -r requirements.in
pdfminer.six==20201018
    # via -r requirements.in
pika==1.2.0
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
protobuf==3.17.3
    # via
    #   google-api-core
    #   googleapis-common-protos
psutil==5.8.0
    # via -r requirements.in
py==1.10.0
    # via pytest
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.20
    # via cffi
pydantic==1.8.2
    # via -r requirements.in
pynacl==1.4.0
    # via paramiko
pyparsing==2.4.7
    # via packaging
pypdf2==1.26.0
    # via -r requirements.in
pyrfc3339==1.1
    # via -r requirements.in
pyrsistent==0.18.0
    # via jsonschema
pytest==6.2.4
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   azure-storage-common
    #   botocore
    #   moto
    #   pandas
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-magic==0.4.24
    # via -r requirements.in
python-redis-lock==3.6.0
    # via -r requirements.in
pytz==2021.1
    # via
    #   django
    #   google-api-core
    #   moto
    #   pandas
    #   pyrfc3339
pyyaml==5.2
    # via
    #   -r requirements.in
    #   awscli
    #   clickclick
    #   connexion
    #   openapi-spec-validator
recordclass==0.15.1
    # via -r requirements.in
redis==3.5.3
    # via
    #   -r requirements.in
    #   fakeredis
    #   python-redis-lock
requests==2.26.0
    # via
    #   -r requirements.in
    #   azure-storage-common
    #   connexion
    #   fhirclient
    #   google-api-core
    #   moto
    #   msrest
    #   requests-oauthlib
    #   responses
requests-oauthlib==1.3.0
    # via msrest
responses==0.13.4
    # via moto
rsa==4.7.2
    # via
    #   awscli
    #   google-auth
    #   python-jose
s3transfer==0.5.0
    # via
    #   awscli
    #   boto3
semantic-version==2.8.5
    # via -r requirements.in
six==1.16.0
    # via
    #   bcrypt
    #   ecdsa
    #   fakeredis
    #   google-api-core
    #   google-auth
    #   grpcio
    #   isodate
    #   jsonschema
    #   nltk
    #   openapi-schema-validator
    #   openapi-spec-validator
    #   protobuf
    #   pynacl
    #   python-dateutil
    #   responses
    #   webtest
slack-sdk==3.8.0
    # via -r requirements.in
smmap==4.0.0
    # via gitdb
sortedcontainers==2.4.0
    # via
    #   fakeredis
    #   pdfminer.six
soupsieve==2.2.1
    # via beautifulsoup4
sqlparse==0.4.1
    # via django
swagger-ui-bundle==0.0.9
    # via connexion
toml==0.10.2
    # via pytest
typing-extensions==3.10.0.0
    # via pydantic
urllib3==1.26.6
    # via
    #   botocore
    #   elastic-apm
    #   requests
    #   responses
waitress==2.0.0
    # via webtest
webob==1.8.7
    # via webtest
webtest==2.0.35
    # via -r requirements.in
werkzeug==1.0.1
    # via
    #   connexion
    #   flask
    #   moto
xmltodict==0.12.0
    # via moto
zipp==3.5.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
